# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FSC Finance App API",
        "description": "API for FSC Finance App that allows users to manage their finances.",
        "version": "1.0.0"
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter 'Bearer' followed by your token. Example: 'Bearer eyuaBASasUUnsn...' "
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
        "/api/users": {
            "post": {
                "summary": "Create a new user",
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "body",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CreateUserParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/UserWithToken"
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "$ref": "#/definitions/InvalidParamResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            },
            "get": {
                "tags": ["User"],
                "summary": "Get authenticated user",
                "responses": {
                    "200": {
                        "description": "User found",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/UserNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete authenticated user",
                "tags": ["User"],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/UserNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "patch": {
                "summary": "Update user by ID",
                "tags": ["User"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "type": "string",
                        "description": "ID of the user to update"
                    },
                    {
                        "name": "body",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CreateUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "$ref": "#/definitions/InvalidParamResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/UserNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "summary": "Login User",
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "body",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/LoginUserParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User logged in successfully",
                        "schema": {
                            "$ref": "#/definitions/UserWithToken"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/refresh-token": {
            "post": {
                "summary": "Refresh User Token",
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "body",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/RefreshTokenParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tokens generated successfully",
                        "schema": {
                            "$ref": "#/definitions/Tokens"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/balance": {
            "get": {
                "summary": "Get authenticated user balance",
                "tags": ["User"],
                "responses": {
                    "200": {
                        "description": "User balance found",
                        "schema": {
                            "$ref": "#/definitions/UserBalance"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/UserNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions": {
            "get": {
                "summary": "Get authenticated user transaction",
                "tags": ["Transaction"],
                "responses": {
                    "200": {
                        "description": "Transactions found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Transaction"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/UserNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new tansaction for the authenticated user",
                "tags": ["Transaction"],
                "responses": {
                    "201": {
                        "description": "Transaction created sucessfully",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "$ref": "#/definitions/InvalidParamResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/UserNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{transactionId}": {
            "patch": {
                "summary": "Update transaction by ID",
                "tags": ["Transaction"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "transactionId",
                        "required": true,
                        "type": "string",
                        "description": "ID of the transaction to update"
                    },
                    {
                        "name": "body",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateTransactionParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction updated sucessfully",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid params",
                        "schema": {
                            "$ref": "#/definitions/InvalidParamResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/TransactionNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete transaction by ID",
                "tags": ["Transaction"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "transactionId",
                        "required": true,
                        "type": "string",
                        "description": "ID of the transaction to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction delete sucessfully",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/TransactionNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "CreateUserParams": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "John",
                    "required": true
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe",
                    "required": true
                },
                "email": {
                    "type": "string",
                    "example": "john@doe.com",
                    "required": true
                },
                "password": {
                    "type": "string",
                    "example": "password",
                    "required": true
                }
            }
        },
        "LoginUserParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@doe.com",
                    "required": true
                },
                "password": {
                    "type": "string",
                    "example": "password",
                    "required": true
                }
            }
        },
        "RefreshTokenParams": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string",
                    "required": true,
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIwMWJjYmRhYy1hZGRmLTRmZDktYWRmNS0zMWI2NzFkMGIzMmMiLCJpYXQiOjE3NTczNTI2NDAsImV4cCI6MTc1OTk0NDY0MH0.YxhZl_tB6QDeBfhuppVgyBMazzJvbg21_LNpG85CkS8"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "30c75525-a1b6-49fb-827b-8e5ab73b4bcf"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john@doe.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "UserWithToken": {
            "allOf": [
                {
                    "$ref": "#/definitions/User"
                },
                {
                    "type": "object",
                    "properties": {
                        "tokens": {
                            "type": "object",
                            "properties": {
                                "accessToken": {
                                    "type": "string",
                                    "example": "eyBASOKJDUmimdia1..."
                                },
                                "refreshToken": {
                                    "type": "string",
                                    "example": "duJIAJIBPADMASIKMdMIASM"
                                }
                            }
                        }
                    }
                }
            ]
        },
        "Tokens": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIwMWJjYmRhYy1hZGRmLTRmZDktYWRmNS0zMWI2NzFkMGIzMmMiLCJpYXQiOjE3NTczNTI2NDAsImV4cCI6MTc1NzM1MzU0MH0.NmijaNHDazTw1Yk-Z4HWT2x9M7eUcWt0dSWksLeP1a6"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIwMWJjYmRhYy1hZGRmLTRmZDktYWRmNS0zMWI2NzFkMGIzMmMiLCJpYXQiOjE3NTczNTI2NDAsImV4cCI6MTc1OTk0NDY0MH0.YxhZl_tB6QDeBfhuppVgyBMazzJvbg21_LNpG85CkS8"
                }
            }
        },
        "UserBalance": {
            "type": "object",
            "properties": {
                "earnings": {
                    "type": "string",
                    "example": 1000
                },
                "expenses": {
                    "type": "string",
                    "example": 500
                },
                "invesments": {
                    "type": "string",
                    "example": 200
                },
                "balance": {
                    "type": "string",
                    "example": 300
                }
            }
        },
        "Transaction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "e5740877-678d-4751-9c0d-6445454aa1d6"
                },
                "user_id": {
                    "type": "string",
                    "example": "30c75525-a1b6-49fb-827b-8e5ab73b4bcf"
                },
                "name": {
                    "type": "string",
                    "example": "Salary"
                },
                "type": {
                    "type": "string",
                    "example": "EARNING",
                    "description": "EARNING, EXPENSE, INVESTMENT"
                },
                "amount": {
                    "type": "string",
                    "example": "1000"
                }
            }
        },
        "CreateTransactionParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Salary",
                    "required": true
                },
                "type": {
                    "type": "string",
                    "example": "EARNING",
                    "description": "EARNING, EXPENSE, INVESTMENT",
                    "required": true
                },
                "date": {
                    "type": "string",
                    "example": "2023-09-16T00:00:00.000Z",
                    "required": true
                },
                "amount": {
                    "type": "number",
                    "example": "1000",
                    "required": true
                }
            }
        },
        "UpdateTransactionParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Salary"
                },
                "type": {
                    "type": "string",
                    "example": "EARNING",
                    "description": "EARNING, EXPENSE, INVESTMENT"
                },
                "date": {
                    "type": "string",
                    "example": "2023-09-16T00:00:00.000Z"
                },
                "amount": {
                    "type": "number",
                    "example": "1000"
                }
            }
        },
        "InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "UserNotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User not found"
                }
            }
        },
        "InvalidParamResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid param (this message depends on the param that is invalid)"
                }
            }
        },
        "TransactionNotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Transaction not found"
                }
            }
        },
        "UnauthorizedErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        }
    }
}